requirements.txt
absl-py==1.4.0
airsim==1.7.0
attrs==23.1.0
boto3==1.28.36
botocore==1.31.36
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
cloudpickle==2.2.1
contourpy==1.1.0
cycler==0.11.0
fastdtw==0.3.4
filelock==3.12.3
fonttools==4.42.1
fsspec==2023.6.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0
grpcio==1.57.0
gym==0.21.0
huggingface-hub==0.16.4
idna==3.4
igraph==0.10.6
imageio==2.31.2
imageio-ffmpeg==0.4.8
importlib-metadata==6.8.0
importlib-resources==6.0.1
jmespath==1.0.1
joblib==1.3.2
kiwisolver==1.4.5
llvmlite==0.40.1
lmdb==1.4.1
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.7.2
msgpack==1.0.4
msgpack-numpy==0.4.8
msgpack-rpc-python==1.0.0
networkx==3.1
numba==0.57.1
numpy==1.24.4
oauthlib==3.2.2
opencv-contrib-python==4.8.0.76
packaging==23.1
pandas==2.0.3
Pillow==10.0.0
protobuf==4.24.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytorch-transformers==1.2.0
pytz==2023.3
PyYAML==6.0.1
regex==2023.8.8
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.6.2
sacremoses==0.0.53
safetensors==0.3.3
seaborn==0.12.2
sentencepiece==0.1.99
six==1.16.0
tensorboard==2.14.0
tensorboard-data-server==0.7.1
tensorboardX==2.6.2.2
texttable==1.6.7
tokenizers==0.13.3
torch
torchaudio
torchvision
tornado==4.5.3
tqdm==4.66.1
transformers==4.32.1
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
Werkzeug==2.3.7
yacs==0.1.8
zipp==3.16.2
